Basic additions
+ Config file
+ Customization
