+++ config.json
@@ -1,0 +1,10 @@
+{
+    "binary": {
+        "0": ".",
+        "1": "_"
+    },
+    "pig-latin": {
+        "case": "lower",
+        "end": "a"
+    }
+}
